requirements.txt
pylint==2.15.*
black==22.10.*
isort==5.10.*

psycopg2-binary

django
django-filter
djangorestframework
djangorestframework-gis
drf-nested-routers
drf-spectacular
factory_boy
